Corrected SVG export - now the text is exported as text, not as paths
Text is properly exported using the SVG <text> elements.
The space occupied by the text is exactly the same as in the original file.
OTOH we loose the information about possible text rotations and about the exact font used.
The produced SVG is much more appropriate for processing and data extraction, but less for the display.